feat: add checks and raise error for null results
In addition added a check all shifts for debug/
admin purposes.

+++ ShiftsLogger.Call911plz/Api/Services/ShiftService.cs
@@ -1,11 +1,12 @@
 
 public interface IShiftService
 {
-    public List<Shift> GetAllShifts(int workerId);
-    public Shift? GetShiftById(int workerId, int id);
+    public List<Shift>? GetShiftsByWorkerId(int workerId);
+    public Shift? GetShiftByShiftId(int workerId, int id);
     public Shift? CreateShift(Shift shift);
     public Shift? UpdateShift(Shift shift);
-    public string? DeleteShift(int id);
+    public string? DeleteShift(int workerId, int id);
+    public List<Shift>? GetAllShifts();
 }
 
 public class ShiftService : IShiftService
@@ -19,7 +20,7 @@ public class ShiftService : IShiftService
 
     public Shift? CreateShift(Shift shift)
     {
-        if (_dbContext.Workers.Find(shift.WorkerId) == null)
+        if (DoesWorkerExists(shift.WorkerId) == false)
             return null;
 
         var savedShift = _dbContext.Shifts.Add(shift);
@@ -27,34 +28,42 @@ public class ShiftService : IShiftService
         return savedShift.Entity;
     }
 
-    public string? DeleteShift(int id)
+    public string? DeleteShift(int workerId, int id)
     {
+        if (DoesWorkerExists(workerId) == false)
+            return null;
+
         Shift? savedShift = _dbContext.Shifts.Find(id);
 
         if (savedShift == null)
             return null;
         
         _dbContext.Shifts.Remove(savedShift);
-        
         _dbContext.SaveChanges();
-
         return $"Successfully deleted shift with id: {id}";
     }
 
-    public List<Shift> GetAllShifts(int workerId)
+    public List<Shift>? GetShiftsByWorkerId(int workerId)
     {
+        if (DoesWorkerExists(workerId) == false)
+            return null;
+        
         return _dbContext.Shifts.Where(shift => shift.WorkerId == workerId).ToList();
     }
 
-    public Shift? GetShiftById(int workerId, int id)
+    public Shift? GetShiftByShiftId(int workerId, int id)
     {
-        Shift? savedShift = _dbContext.Shifts.Find(workerId, id);
+        if (DoesWorkerExists(workerId) == false)
+            return null;
+
+        // Does not need a shift check as it does not alter db
+        Shift? savedShift = _dbContext.Shifts.Find(id);
         return savedShift;
     }
 
     public Shift? UpdateShift(Shift shift)
     {
-        if (_dbContext.Workers.Find(shift.WorkerId) == null)
+        if (DoesWorkerExists(shift.WorkerId) == false)
             return null;
         
         Shift? savedShift = _dbContext.Shifts.Find(shift.Id);
@@ -64,7 +73,18 @@ public class ShiftService : IShiftService
 
         _dbContext.Entry(savedShift).CurrentValues.SetValues(shift);
         _dbContext.SaveChanges();
-
         return savedShift;
+    }
+
+    // Debug/Admin purposes only
+    public List<Shift>? GetAllShifts()
+    {
+        List<Shift>? shifts = _dbContext.Shifts.ToList();
+        return shifts;
+    }
+
+    private bool DoesWorkerExists(int workerId)
+    {
+        return _dbContext.Workers.Where(worker => worker.EmployeeId == workerId).Any();
     }
 }
